services/lesson_generator/requirements.txt
# 2025-10-22: Dependencies for Lesson JSON Generator Microservice
# Authors: Cascade

Flask==3.0.0
openai==1.3.3
python-dotenv==1.0.0
